:recycle: 도메인 로직 useCourseMap 훅으로 분리
